+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Using payments flags from `vtex.playment-flags`
+- Get PaymentMethods from `userPaymentMethods`
+
 ## [1.8.1] - 2019-09-26
 
 ### Changed